Rename utils to _clean
